move testingService to node-core
