+++ packages/node-core/src/configure/index.ts
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './NodeConfig';
+export * from './SubqueryProject';
